moved members screen  to screens folder
