+++ lib/routes/routes.dart
@@ -2,8 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:paradox/screens/Referral.dart';
 import 'package:paradox/screens/stageCompleted_screen.dart';
 import 'package:paradox/screens/stats_screen.dart';
-import '../utilities/member_screen.dart';
-import 'package:paradox/utilities/member_screen.dart';
+import 'package:paradox/screens/member_screen.dart';
 import '../screens/question_screen.dart';
 import '../screens/leaderboard_screen.dart';
 import 'package:paradox/screens/leaderboard_screen.dart';
